refactor(mux): privatize doctor log forensics
Move bounded mux log scanning and diagnosis behind one binary-private collector so doctor-only DTOs no longer escape the library.
