+++ crates/rimz/src/mux/tmux/tests.rs
@@ -158,43 +158,6 @@ fn version_parser_and_floor_hold() {
 }
 
 #[test]
-fn log_classifier_matches_error_and_fatal_mentions() {
-    use crate::mux::logtail::LogSeverity;
-
-    assert_eq!(
-        classify_log_line("server error: client lost"),
-        Some(LogSeverity::Error)
-    );
-    assert_eq!(
-        classify_log_line("fatal: control socket closed"),
-        Some(LogSeverity::Error)
-    );
-    assert_eq!(
-        classify_log_line("server panic: invariant failed"),
-        Some(LogSeverity::Panic)
-    );
-    assert_eq!(classify_log_line("normal redraw"), None);
-}
-
-#[test]
-fn log_lines_carry_the_epoch_stamp_tmux_opens_them_with() {
-    use crate::mux::logtail::RecordLine;
-
-    let stamped = |line: &str| match parse_log_line(line) {
-        RecordLine::Start(start) => start.at,
-        RecordLine::Continuation => panic!("record start"),
-    };
-
-    assert_eq!(
-        stamped("1784493802.501234 server error: client lost"),
-        Some(jiff::Timestamp::new(1_784_493_802, 501_234_000).unwrap())
-    );
-    // A line tmux did not stamp survives every `--clear` cutoff rather than
-    // being dismissed on a guess.
-    assert_eq!(stamped("server error: no stamp here"), None);
-}
-
-#[test]
 fn tmux_extended_key_bindings_follow_extended_key_format() {
     let csi_u = crate::config::TmuxConfig {
         extended_keys_format: crate::config::TmuxExtendedKeysFormat::CsiU,
