+++ crates/rimz/src/mux/zellij.rs
@@ -143,324 +143,6 @@ pub fn live_presence_plugin_ids(session_name: &str) -> Result<Vec<u32>> {
     ZellijBackend::new().live_presence_plugin_ids(session_name)
 }
 
-#[cfg(test)]
-fn classify_log_line(line: &str) -> Option<super::logtail::LogSeverity> {
-    match parse_log_line(line) {
-        super::logtail::RecordLine::Start(start) => start.severity,
-        super::logtail::RecordLine::Continuation => None,
-    }
-}
-
-pub fn parse_log_line(line: &str) -> super::logtail::RecordLine {
-    use super::logtail::{LogRecordStart, LogSeverity, RecordLine};
-
-    if line.starts_with("Panic occured") || line.starts_with("Panic occurred") {
-        return RecordLine::Start(LogRecordStart {
-            severity: Some(LogSeverity::Panic),
-            message: line.to_owned(),
-            ..LogRecordStart::default()
-        });
-    }
-    let Some((severity_name, rest)) = ["TRACE", "DEBUG", "INFO", "WARN", "ERROR"]
-        .into_iter()
-        .find_map(|severity| {
-            line.strip_prefix(severity)
-                .filter(|rest| rest.chars().next().is_none_or(char::is_whitespace))
-                .map(|rest| (severity, rest))
-        })
-    else {
-        return RecordLine::Continuation;
-    };
-    let mut severity = match severity_name {
-        "WARN" => Some(LogSeverity::Warn),
-        "ERROR" => Some(LogSeverity::Error),
-        _ => None,
-    };
-    if let Some(header) = parse_zellij_structured_header(rest) {
-        if header.message.starts_with("Panic occured")
-            || header.message.starts_with("Panic occurred")
-        {
-            severity = Some(LogSeverity::Panic);
-        }
-        return RecordLine::Start(LogRecordStart {
-            severity,
-            at: parse_log_timestamp(&header.timestamp),
-            target: Some(header.target),
-            thread: Some(header.thread),
-            source: Some(header.source),
-            message: header.message,
-        });
-    }
-    let message = rest.trim_start().to_owned();
-    RecordLine::Start(LogRecordStart {
-        severity,
-        message,
-        ..LogRecordStart::default()
-    })
-}
-
-/// Zellij stamps each record with local wall-clock time and no offset
-/// (`2026-07-19 13:37:49.089`), so the machine's own zone resolves it.
-fn parse_log_timestamp(raw: &str) -> Option<jiff::Timestamp> {
-    raw.replace(' ', "T")
-        .parse::<jiff::civil::DateTime>()
-        .ok()?
-        .to_zoned(jiff::tz::TimeZone::system())
-        .ok()
-        .map(|zoned| zoned.timestamp())
-}
-
-struct ZellijLogHeader {
-    target: String,
-    timestamp: String,
-    thread: String,
-    source: String,
-    message: String,
-}
-
-fn parse_zellij_structured_header(rest: &str) -> Option<ZellijLogHeader> {
-    let rest = rest.trim_start().strip_prefix('|')?;
-    let (target, rest) = rest.split_once('|')?;
-    let (timestamp, rest) = rest.trim_start().split_once(" [")?;
-    let (thread, rest) = rest.split_once(']')?;
-    let (source, message) = rest.trim_start().split_once(": ")?;
-    let target = target.trim();
-    let timestamp = timestamp.trim();
-    let thread = thread.trim();
-    let source = source.trim();
-    if target.is_empty() || timestamp.is_empty() || thread.is_empty() || source.is_empty() {
-        return None;
-    }
-    Some(ZellijLogHeader {
-        target: target.to_owned(),
-        timestamp: timestamp.to_owned(),
-        thread: thread.to_owned(),
-        source: source.to_owned(),
-        message: message.trim_end().to_owned(),
-    })
-}
-
-/// The wrapper zellij prints above every recoverable failure; it names nothing
-/// on its own, so the `Caused by:` chain underneath is the real subject.
-const NON_FATAL_HEADER: &str = "a non-fatal error occured";
-
-pub fn diagnose_log_record(
-    previous: Option<&super::logtail::LogicalRecord>,
-    record: &super::logtail::LogicalRecord,
-    next: Option<&super::logtail::LogicalRecord>,
-) -> Option<super::logtail::LogDiagnosis> {
-    use super::logtail::{LogDiagnosis, LogImpact, LogSeverity, LogState, normalized_issue_key};
-
-    let severity = record.start.severity?;
-    // A disconnect writes two records; the second rides with the first.
-    if previous.is_some_and(is_unknown_client_message) && is_client_send_failure(record) {
-        return None;
-    }
-    let paired_send_failure =
-        next.filter(|next| is_unknown_client_message(record) && is_client_send_failure(next));
-
-    let target = record.start.target.as_deref().unwrap_or_default();
-    let message = record.start.message.trim();
-    let causes = record_causes(&record.text);
-    let subject = match (message.starts_with(NON_FATAL_HEADER), causes.first()) {
-        (true, Some(cause)) => cause.as_str(),
-        _ => message,
-    };
-
-    if let Some(expected) = expected_lifecycle(record, subject, paired_send_failure) {
-        return Some(expected);
-    }
-
-    // The sidebar reads panes through these plugin calls, so a timeout here is
-    // the log's own account of pane discovery falling behind.
-    if subject.contains("timed out") && subject.contains("for plugin") {
-        return Some(LogDiagnosis {
-            key: "plugin_pane_query_timeout".to_owned(),
-            state: LogState::Investigate,
-            impact: LogImpact::Warn,
-            summary: "plugin pane queries timed out — pane discovery lags behind the room"
-                .to_owned(),
-            sample: None,
-        });
-    }
-    // An unknown client message with no disconnect behind it, and the logout
-    // zellij escalates to, are the same event stream: a client speaking a
-    // protocol this server does not know.
-    if is_unknown_client_message(record)
-        || (message.starts_with("Client sent over") && message.contains("unknown messages"))
-    {
-        return Some(LogDiagnosis {
-            key: "client_protocol_mismatch".to_owned(),
-            state: LogState::Investigate,
-            impact: LogImpact::Warn,
-            summary: "a client sent messages zellij could not read — usually a client/server version mismatch"
-                .to_owned(),
-            sample: None,
-        });
-    }
-
-    // Zellij keeps the pane and spawns it in the inherited directory, so the
-    // pane lives and only its directory is wrong. Keying on the path keeps two
-    // different stale directories in two groups, each naming its own fix.
-    if let Some(cwd) = missing_pane_cwd(subject) {
-        return Some(LogDiagnosis {
-            key: normalized_issue_key(&format!("missing_pane_cwd:{cwd}")),
-            state: LogState::Investigate,
-            impact: LogImpact::Warn,
-            summary: format!(
-                "a pane's configured directory is missing ({cwd}) — zellij started it in the inherited directory"
-            ),
-            sample: None,
-        });
-    }
-
-    let impact = match severity {
-        LogSeverity::Warn => LogImpact::Warn,
-        LogSeverity::Error | LogSeverity::Panic => LogImpact::Alarm,
-    };
-    // Naming the whole cause chain keeps unrelated failures in separate groups;
-    // keyed on the wrapper alone they collapse into one meaningless bucket.
-    let summary = if causes.is_empty() || !message.starts_with(NON_FATAL_HEADER) {
-        message.to_owned()
-    } else {
-        causes.join(": ")
-    };
-    Some(LogDiagnosis {
-        key: normalized_issue_key(&format!("{target}:{summary}")),
-        state: LogState::Investigate,
-        impact,
-        summary,
-        sample: None,
-    })
-}
-
-/// Log traffic the room provokes by living its normal life: clients attaching
-/// and leaving, panes closing, a busy server acknowledging late. Each one reads
-/// as an ERROR in zellij's log and means nothing to the operator.
-fn expected_lifecycle(
-    record: &super::logtail::LogicalRecord,
-    subject: &str,
-    paired_send_failure: Option<&super::logtail::LogicalRecord>,
-) -> Option<super::logtail::LogDiagnosis> {
-    use super::logtail::{LogDiagnosis, LogImpact, LogState};
-
-    let expected = |key: &str, summary: &str, sample: Option<String>| LogDiagnosis {
-        key: key.to_owned(),
-        state: LogState::Expected,
-        impact: LogImpact::Info,
-        summary: summary.to_owned(),
-        sample,
-    };
-
-    // Only the proven pair reads as a departure: an unknown client message on
-    // its own is evidence of something else, and gets to keep saying so.
-    if paired_send_failure.is_some() || is_client_send_failure(record) {
-        return Some(expected(
-            "client_disconnect",
-            "a client left the session",
-            paired_send_failure.map(|next| format!("{}\n{}", record.text, next.text)),
-        ));
-    }
-    if let Some(action) = action_ack_timeout(subject) {
-        return Some(expected(
-            &format!("action_ack_timeout:{action}"),
-            &format!("zellij acknowledged {action} late (the action still ran)"),
-            None,
-        ));
-    }
-    // Zellij truncates the target column, so the untruncated source path is the
-    // reliable way to place a record in the server's pty reader.
-    let source = record.start.source.as_deref().unwrap_or_default();
-    if source.contains("terminal_bytes.rs") && subject.contains("I/O error (os error 5)") {
-        return Some(expected(
-            "closed_pane_pty",
-            "read from a closed pane's terminal",
-            None,
-        ));
-    }
-    if subject.starts_with("failed to disable mouse mode") {
-        return Some(expected(
-            "client_teardown_mouse_mode",
-            "a client tore down mouse mode on a terminal already gone",
-            None,
-        ));
-    }
-    // Pane-targeting actions name a pane the room listed a moment earlier, so a
-    // pane that closes inside that window resolves to nothing. The id varies per
-    // occurrence and one key groups them, because the race is the single fact.
-    if subject.starts_with("Pane with id") && subject.ends_with("not found") {
-        return Some(expected(
-            "closed_pane_action",
-            "addressed a pane that had already closed",
-            None,
-        ));
-    }
-    let lower = record.text.to_ascii_lowercase();
-    if lower.contains("closed terminal") && lower.contains("resize") && lower.contains("caused by")
-    {
-        return Some(expected(
-            "closed_terminal_resize",
-            "resized a pane whose terminal had closed",
-            None,
-        ));
-    }
-    None
-}
-
-/// The directory a pane asked for and zellij could not enter, from
-/// `Failed to set CWD for new pane. '<path>' does not exist or is not a folder`.
-/// Matching the whole wording keeps a reworded upstream message falling through
-/// to the generic path rather than reporting a truncated directory.
-fn missing_pane_cwd(subject: &str) -> Option<&str> {
-    subject
-        .strip_prefix("Failed to set CWD for new pane. '")?
-        .strip_suffix("' does not exist or is not a folder")
-}
-
-/// The action zellij took too long to acknowledge, from
-/// `Action CliPipe did not complete within 1s timeout`.
-fn action_ack_timeout(subject: &str) -> Option<&str> {
-    subject
-        .strip_prefix("Action ")?
-        .split_once(" did not complete within")
-        .map(|(action, _)| action)
-}
-
-fn is_unknown_client_message(record: &super::logtail::LogicalRecord) -> bool {
-    record.start.message == "Received unknown message from client."
-}
-
-fn is_client_send_failure(record: &super::logtail::LogicalRecord) -> bool {
-    record.start.message.starts_with(NON_FATAL_HEADER)
-        && record.text.contains("failed to send message to client")
-        && record.text.contains("Broken pipe (os error 32)")
-}
-
-/// The `Caused by:` chain under an error record, outermost cause first. Anyhow
-/// numbers the entries once there is more than one; a lone cause is bare.
-fn record_causes(text: &str) -> Vec<String> {
-    text.lines()
-        .skip_while(|line| line.trim() != "Caused by:")
-        .skip(1)
-        .map(str::trim)
-        .take_while(|line| !line.is_empty())
-        .map(|line| strip_cause_index(line).trim().to_owned())
-        .collect()
-}
-
-/// Drop anyhow's `0: ` ordinal, keeping the cause text itself.
-fn strip_cause_index(line: &str) -> &str {
-    line.split_once(' ')
-        .filter(|(ordinal, _)| {
-            ordinal.ends_with(':')
-                && ordinal
-                    .trim_end_matches(':')
-                    .chars()
-                    .all(|ch| ch.is_ascii_digit())
-        })
-        .map_or(line, |(_, rest)| rest)
-}
-
 /// Parse `"zellij 0.41.2"` (and tolerant of leading/trailing whitespace).
 /// Returns None when the shape is unexpected so `doctor` can render the raw
 /// string verbatim.
