+++ crates/rimz/src/mux/zellij/tests.rs
@@ -1222,253 +1222,6 @@ fn version_parser_accepts_zellij_output_shapes() {
 }
 
 #[test]
-fn log_classifier_matches_zellij_levels() {
-    use crate::mux::logtail::LogSeverity;
-
-    for (line, expected) in [
-        ("Panic occured: unknown messages", Some(LogSeverity::Panic)),
-        ("Panic occurred: unknown messages", Some(LogSeverity::Panic)),
-        ("ERROR failed to decode", Some(LogSeverity::Error)),
-        (
-            "ERROR  |zellij_utils::errors::not| 2026-07-17 04:06:02.158 [screen] zellij-utils/src/errors.rs:819: Panic occured:",
-            Some(LogSeverity::Panic),
-        ),
-        ("WARN slow client", Some(LogSeverity::Warn)),
-        ("INFO later WARN text is not a level", None),
-        ("WARNING is not WARN token", None),
-    ] {
-        assert_eq!(classify_log_line(line), expected, "{line}");
-    }
-}
-
-#[test]
-fn log_diagnosis_names_a_wrapped_error_by_its_cause() {
-    use crate::mux::logtail::{LogImpact, LogState, LogicalRecord, RecordLine};
-
-    // Zellij prints the same wrapper above every recoverable failure, so two
-    // unrelated failures share a header and differ only underneath it.
-    let record = |causes: &str| {
-        let line = "ERROR  |???                      | 2026-07-17 12:23:34.169 [unnamed] zellij-client/src/lib.rs:975: a non-fatal error occured";
-        let RecordLine::Start(start) = parse_log_line(line) else {
-            panic!("record start");
-        };
-        LogicalRecord {
-            start,
-            text: format!("{line}\n\nCaused by:\n{causes}"),
-            truncated: false,
-        }
-    };
-    let mouse = record("    0: failed to set the cursor shape\n    1: I/O error (os error 5)");
-    let write = record("    0: failed to write to the pty\n    1: I/O error (os error 5)");
-
-    let mouse = diagnose_log_record(None, &mouse, None).unwrap();
-    let write = diagnose_log_record(None, &write, None).unwrap();
-
-    assert_eq!(
-        mouse.summary,
-        "failed to set the cursor shape: I/O error (os error 5)"
-    );
-    assert_eq!(mouse.state, LogState::Investigate);
-    assert_eq!(mouse.impact, LogImpact::Alarm);
-    assert_ne!(
-        mouse.key, write.key,
-        "two failures under one wrapper stay two issues"
-    );
-}
-
-#[test]
-fn log_diagnosis_requires_complete_known_lifecycle_evidence() {
-    use crate::mux::logtail::{LogImpact, LogState, LogicalRecord};
-
-    let unknown_line = "ERROR  |zellij_server::route     | 2026-07-17 12:23:34.169 [server_router] zellij-server/src/route.rs:2642: Received unknown message from client.";
-    let unknown_start = match parse_log_line(unknown_line) {
-        crate::mux::logtail::RecordLine::Start(start) => start,
-        crate::mux::logtail::RecordLine::Continuation => panic!("record start"),
-    };
-    assert_eq!(
-        unknown_start.target.as_deref(),
-        Some("zellij_server::route")
-    );
-    assert!(
-        unknown_start.at.is_some(),
-        "the structured header carries a readable time"
-    );
-    assert_eq!(unknown_start.thread.as_deref(), Some("server_router"));
-    assert_eq!(
-        unknown_start.source.as_deref(),
-        Some("zellij-server/src/route.rs:2642")
-    );
-    let broken_line = "ERROR  |???                      | 2026-07-17 12:23:34.169 [unnamed] zellij-server/src/os_input_output.rs:231: a non-fatal error occured";
-    let broken_start = match parse_log_line(broken_line) {
-        crate::mux::logtail::RecordLine::Start(start) => start,
-        crate::mux::logtail::RecordLine::Continuation => panic!("record start"),
-    };
-    let unknown = LogicalRecord {
-        start: unknown_start,
-        text: unknown_line.to_owned(),
-        truncated: false,
-    };
-    let broken_pipe = LogicalRecord {
-        start: broken_start,
-        text: format!(
-            "{broken_line}\n\nCaused by:\n    0: failed to send message to client 2\n    1: Broken pipe (os error 32)"
-        ),
-        truncated: false,
-    };
-
-    let expected = diagnose_log_record(None, &unknown, Some(&broken_pipe)).unwrap();
-    assert_eq!(expected.state, LogState::Expected);
-    assert_eq!(expected.impact, LogImpact::Info);
-    assert!(expected.sample.unwrap().contains("Broken pipe"));
-    assert!(diagnose_log_record(Some(&unknown), &broken_pipe, None).is_none());
-    let investigate = diagnose_log_record(None, &unknown, None).unwrap();
-    assert_eq!(investigate.state, LogState::Investigate);
-    assert_eq!(investigate.impact, LogImpact::Warn);
-    assert!(
-        investigate.summary.contains("version mismatch"),
-        "an unpaired unknown message names what it usually means: {}",
-        investigate.summary
-    );
-
-    let cli_pipe = match parse_log_line(
-        "ERROR  |zellij_server::route| 2026-07-17 12:23:44.875 [server_router] zellij-server/src/route.rs:75: Action CliPipe did not complete within 1s timeout",
-    ) {
-        crate::mux::logtail::RecordLine::Start(start) => LogicalRecord {
-            text: start.message.clone(),
-            start,
-            truncated: false,
-        },
-        crate::mux::logtail::RecordLine::Continuation => panic!("record start"),
-    };
-    assert_eq!(
-        diagnose_log_record(None, &cli_pipe, None).unwrap().state,
-        LogState::Expected
-    );
-}
-
-/// Both records are zellij ERRORs the room provokes on its own, and neither
-/// costs the reader a pane: a pane-targeting action can always lose its target
-/// to a close, and a stale directory still yields a live pane in the inherited
-/// one. Reporting either as an alarm spends the reader's attention on nothing.
-#[test]
-fn log_diagnosis_grades_self_inflicted_pane_errors_below_alarm() {
-    use crate::mux::logtail::{LogImpact, LogState, LogicalRecord, RecordLine};
-
-    let record = |line: &str| {
-        let RecordLine::Start(start) = parse_log_line(line) else {
-            panic!("record start");
-        };
-        LogicalRecord {
-            text: start.message.clone(),
-            start,
-            truncated: false,
-        }
-    };
-
-    let closed = record(
-        "ERROR  |zellij_server::screen    | 2026-07-20 00:02:56.758 [screen] zellij-server/src/screen.rs:9730: Pane with id Terminal(336) not found",
-    );
-    let closed = diagnose_log_record(None, &closed, None).unwrap();
-    assert_eq!(closed.state, LogState::Expected);
-    assert_eq!(closed.impact, LogImpact::Info);
-
-    // The id varies per occurrence; one key keeps the race a single issue.
-    let other = record(
-        "ERROR  |zellij_server::screen    | 2026-07-20 00:02:57.758 [screen] zellij-server/src/screen.rs:9730: Pane with id Terminal(412) not found",
-    );
-    assert_eq!(
-        closed.key,
-        diagnose_log_record(None, &other, None).unwrap().key,
-    );
-
-    let cwd = record(
-        "ERROR  |zellij_server::os_input_o| 2026-07-19 23:28:54.038 [pty] zellij-server/src/os_input_output_unix.rs:216: Failed to set CWD for new pane. '/tmp/rimz-presence-probe' does not exist or is not a folder",
-    );
-    let cwd = diagnose_log_record(None, &cwd, None).unwrap();
-    assert_eq!(cwd.state, LogState::Investigate);
-    assert_eq!(cwd.impact, LogImpact::Warn);
-    assert!(
-        cwd.summary.contains("/tmp/rimz-presence-probe"),
-        "the directory to fix is the whole point of the line: {}",
-        cwd.summary
-    );
-
-    // Two stale directories are two fixes, so they stay two issues.
-    let elsewhere = record(
-        "ERROR  |zellij_server::os_input_o| 2026-07-19 23:28:55.038 [pty] zellij-server/src/os_input_output_unix.rs:216: Failed to set CWD for new pane. '/tmp/gone' does not exist or is not a folder",
-    );
-    assert_ne!(
-        cwd.key,
-        diagnose_log_record(None, &elsewhere, None).unwrap().key,
-    );
-
-    // A reworded upstream message keeps its alarm rather than reporting a
-    // truncated directory.
-    let reworded = record(
-        "ERROR  |zellij_server::os_input_o| 2026-07-19 23:28:56.038 [pty] zellij-server/src/os_input_output_unix.rs:216: Failed to set CWD for new pane. '/tmp/gone' is unreadable",
-    );
-    assert_eq!(
-        diagnose_log_record(None, &reworded, None).unwrap().impact,
-        LogImpact::Alarm,
-    );
-}
-
-#[test]
-fn logical_log_scan_groups_complete_0443_artifacts_conservatively() {
-    let dir = tempfile::tempdir().unwrap();
-    let path = dir.path().join("zellij.log");
-    std::fs::write(
-        &path,
-        concat!(
-            "ERROR  |zellij_server::route     | 2026-07-17 12:23:34.169 [server_router] zellij-server/src/route.rs:2642: Received unknown message from client.\n",
-            "ERROR  |???                      | 2026-07-17 12:23:34.169 [unnamed] zellij-server/src/os_input_output.rs:231: a non-fatal error occured\n",
-            "\nCaused by:\n    0: failed to send message to client 2\n    1: Broken pipe (os error 32)\n",
-            "INFO   |zellij_server            | 2026-07-17 12:23:35.000 [main] zellij-server/src/lib.rs:1: healthy\n",
-            "ERROR  |zellij_server::route     | 2026-07-17 12:23:44.875 [server_router] zellij-server/src/route.rs:75: Action CliPipe did not complete within 1s timeout\n",
-            "ERROR  |zellij_server::route     | 2026-07-17 12:23:45.875 [server_router] zellij-server/src/route.rs:75: Action CliPipe did not complete within 1s timeout\n",
-            "ERROR  |zellij_server::pty       | 2026-07-17 12:23:46.000 [pty] zellij-server/src/pty.rs:9: pane query failed\n",
-            "ERROR  |zellij_utils::errors::not| 2026-07-17 12:23:47.000 [screen] zellij-utils/src/errors.rs:819: Panic occurred:\n",
-            "    thread: screen\n    message: fatal\n",
-        ),
-    )
-    .unwrap();
-
-    let scan = crate::mux::logtail::scan_tail(
-        &path,
-        crate::mux::logtail::LogWindow {
-            bytes: 64 * 1024,
-            issue_cap: 10,
-            since: None,
-        },
-        parse_log_line,
-        diagnose_log_record,
-    )
-    .unwrap();
-
-    assert_eq!(scan.logical_records, 7);
-    assert_eq!(scan.problem_records, 5);
-    assert_eq!(scan.issues.len(), 4);
-    assert_eq!(
-        scan.issues[0].state,
-        crate::mux::logtail::LogState::Expected
-    );
-    assert!(scan.issues[0].samples[0].contains("Broken pipe"));
-    assert_eq!(scan.issues[1].occurrences, 2);
-    assert_eq!(
-        scan.issues[1].state,
-        crate::mux::logtail::LogState::Expected
-    );
-    assert_eq!(
-        scan.issues[2].state,
-        crate::mux::logtail::LogState::Investigate
-    );
-    assert_eq!(
-        scan.issues[3].severity,
-        crate::mux::logtail::LogSeverity::Panic
-    );
-}
-
-#[test]
 fn version_serves_the_memoized_probe() {
     let backend = ZellijBackend::default();
     backend
