+++ refactor-target.toml
@@ -231,7 +231,7 @@ allowed-imports = [
     "store",
     "workspace",
 ]
-surface-budget = 262
+surface-budget = 245
 
 [[module]]
 path = "crates/rimz/src/mux/mount_proof.rs"
@@ -272,6 +272,11 @@ surface-budget = 15
 [[strangler]]
 symbol = "purge_zellij_session_cache_in"
 path = "crates/rimz/src/mux/recovery.rs"
+baseline = 0
+
+[[strangler]]
+symbol = "logtail"
+path = "crates/rimz/src/mux"
 baseline = 0
 
 [[module]]
